fix: Move files properly to make tsup build properly.
