+++ packages/utilities/src/index.ts
@@ -2,5 +2,4 @@
 export { useDefineTemplateAndProjector } from "./useDefineTemplateAndProjector.js"
 export * from "./lib/helpers/conditional"
 export * from "./lib/helpers/get-collection"
-export * from "./lib/helpers/throw"
 export * from "./lib/helpers"
