+++ packages/utilities/src/lib/helpers/conditional.ts
@@ -54,4 +54,19 @@ export function executeUnless<T extends Callback>(condition: boolean, cb: T) {
 
 }
 
+export function throwIf(condition: boolean, message = "Something went wrong", cause?: unknown): asserts condition is false {
 
+
+    executeIf(condition, () => {
+        throw new Error(message, { cause })
+    })
+
+
+}
+
+
+export function throwUnless(condition: boolean, message = "Something went wrong", cause?: unknown): asserts condition {
+
+    throwIf(!condition, message, cause)
+
+}
