--- packages/utilities/src/lib/index.ts
@@ -1,2 +1,0 @@
-export * from "packages/utilities/src/lib/helpers"
-export * from "packages/utilities/src/lib/types"
