+++ packages/utilities/tsup.config.ts
@@ -2,8 +2,11 @@ import { defineConfig } from 'tsup'
 
 export default defineConfig((ctx) => ({
     entry: ['src/index.ts'],
-    publicDir: "src/lib/components",
+    publicDir: true,
     splitting: false,
+    external: [
+        "astro:content"
+    ],
     format: ["cjs", "esm"],
     dts: true,
     minify: !ctx.watch,
